fixed auth

+++ app/src/main/java/com/example/burgerapplication/dto/User.kt
@@ -4,5 +4,5 @@ data class User(
     val username: String,
     val firstName: String,
     val lastName: String,
-    val points: Int
+    val points: Double
 )
